chore: update .vulture

+++ migrations/001_initial_schema.sql
@@ -201,7 +201,7 @@ TTL created_at + INTERVAL 1 YEAR;
 CREATE TABLE IF NOT EXISTS valueData (
     series_id UInt32,
     timestamp DateTime64(6),
-    value Float64,
+    value Decimal(18, 6),
     created_at DateTime64(6) DEFAULT now64(6),
     updated_at DateTime64(6) DEFAULT now64(6)
 ) ENGINE = MergeTree
